+++ kubejs/server_scripts/alekiships.js
@@ -1,0 +1,7 @@
+
+
+const AlekishipsRecipes = (event) => {
+  //Removal
+    event.remove({id: 'alekiships:crafting/cannon'})
+    event.remove({id: 'alekiships:crafting/anchor'})
+}