+++ kubejs/server_scripts/immersiveengineering.js
@@ -11,6 +11,10 @@ const ImmersiveEngineeringRecipes = (event) => {
     event.remove({output: 'immersiveengineering:ingot_silver'})
     event.remove({output: 'immersiveengineering:ingot_steel'})
     event.remove({id: 'immersiveengineering:blastfurnace/steel_block'})
+    event.remove({id: 'immersiveengineering:crafting/storage_constantan_to_ingot_constantan'})
+    event.remove({id: 'immersiveengineering:crafting/storage_electrum_to_ingot_electrum'})
+    event.remove({id: 'immersiveengineering:crafting/storage_aluminum_to_ingot_aluminum'})
+    event.remove({id: 'immersiveengineering:crafting/storage_lead_to_ingot_lead'})
 
     //replace
     event.replaceInput(
@@ -27,6 +31,25 @@ const ImmersiveEngineeringRecipes = (event) => {
         {id: 'tfc_ie_addon:metalpress/uranium_block'},
         '#forge:ingots/uranium',
         Item.of('tfc_ie_addon:metal/double_ingot/uranium', 1)
+    )
+    //IE Press
+    event.recipes.immersiveengineering.metal_press(
+        'immersiveengineering:storage_lead', 'tfc_ie_addon:metal/double_ingot/lead', 'tfc_ie_addon:mold_block', 1000
+    )
+    event.recipes.immersiveengineering.metal_press(
+        'immersiveengineering:storage_electrum', 'tfc_ie_addon:metal/double_ingot/electrum', 'tfc_ie_addon:mold_block', 1000
+    )
+    event.recipes.immersiveengineering.metal_press(
+        'immersiveengineering:storage_nickel', 'tfc:metal/double_ingot/nickel', 'tfc_ie_addon:mold_block', 1000
+    )
+    event.recipes.immersiveengineering.metal_press(
+        'immersiveengineering:storage_silver', 'tfc:metal/double_ingot/silver', 'tfc_ie_addon:mold_block', 1000
+    )
+    event.recipes.immersiveengineering.metal_press(
+        'immersiveengineering:storage_constantan', 'tfc_ie_addon:metal/double_ingot/constantan', 'tfc_ie_addon:mold_block', 1000
+    )
+    event.recipes.immersiveengineering.metal_press(
+        'immersiveengineering:storage_aluminum', 'tfc_ie_addon:metal/double_ingot/aluminum', 'tfc_ie_addon:mold_block', 1000
     )
 
     //TFC Heating